// js_modules/dagit/packages/core/src/runs/types/ComputeLogsSubscriptionFragment.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL fragment: ComputeLogsSubscriptionFragment
// ====================================================

export interface ComputeLogsSubscriptionFragment {
  __typename: "ComputeLogFile";
  data: string | null;
  cursor: number;
  path: string;
  downloadUrl: string | null;
}